requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.3
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-extensions==3.2.1
django-ipware==4.0.2
djangorestframework==3.14.0
frozenlist==1.3.3
geoip2==4.6.0
idna==3.4
maxminddb==2.2.0
multidict==6.0.3
mysqlclient==2.1.1
pymemcache==4.0.0
pyngrok==5.2.1
python-memcached==1.59
pytz==2022.6
PyYAML==6.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.13
yarl==1.8.2
